extract remaining routes in misc controller
